backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
huggingface-hub==0.34.4
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
pillow==11.3.0
pydantic-settings==2.10.1
python-multipart==0.0.20
transformers==4.55.4
uvicorn[standard]==0.35.0
